emergent-microservices/verifiable-reality/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
gunicorn==21.2.0
pydantic==2.6.4
pydantic-settings==2.2.1

# Database drivers
# Security
PyJWT[crypto]==2.8.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5

# Observability
prometheus-client==0.20.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
httpx==0.27.0

# Development
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
ruff==0.3.4
mypy==1.9.0
black==24.3.0
isort==5.13.2
bandit==1.7.8
hypothesis==6.100.0
